feat: move message compatibility test to utils
