+++ y-octo-utils/Cargo.toml
@@ -9,14 +9,19 @@ version = "0.0.1"
 
 [features]
 bench = ["regex"]
-fuzz  = ["arbitrary", "phf", "y-octo", "yrs"]
+fuzz  = ["arbitrary", "phf"]
 
 [dependencies]
 arbitrary = { version = "1.3.0", features = ["derive"], optional = true }
 phf       = { version = "0.11", features = ["macros"], optional = true }
 regex     = { version = "1.5", optional = true }
-y-octo    = { workspace = true, optional = true }
-yrs       = { version = "=0.16.5", optional = true }
+y-octo    = { workspace = true }
+y-sync    = "=0.3.0"
+yrs       = "=0.16.5"
+
+[dev-dependencies]
+proptest         = "1.2"
+proptest-derive  = "0.4"
 
 [[bin]]
 name = "bench_result_render"
