+++ y-octo-utils/src/lib.rs
@@ -1,5 +1,9 @@
+mod message;
+
 #[cfg(feature = "fuzz")]
 pub mod doc_operation;
 
 #[cfg(feature = "fuzz")]
 pub use doc_operation::*;
+
+pub use message::{to_sync_message, to_y_message};
